[lint] refactor unittestcases.py
splitted unittestcases.py in

* unittestcases.py
* archiveobscpiotestcases.py

to get rid of 'too much public methods error'
