+++ tests/fixtures/ArchiveOBSCpioTestCases/test_obscpio_extract_d
